+++ DogSitterMarketplaceBll/Models/Orders/Request/OrderUpdate.cs
@@ -17,7 +17,7 @@ namespace DogSitterMarketplaceBll.Models.Orders.Request
 
         public int SitterWorkId { get; set; }
 
-        public int Summ { get; set; }
+        public decimal Summ { get; set; }
 
         public DateTime DateStart { get; set; }
