+++ DogSitterMarketplaceBll/Models/Orders/Response/OrderResponse.cs
@@ -16,7 +16,7 @@ namespace DogSitterMarketplaceBll.Models.Orders.Response
 
         public SitterWorkResponse SitterWork { get; set; }
 
-        public int Summ { get; set; }
+        public decimal Summ { get; set; }
 
         public DateTime DateStart { get; set; }
